pyproject.toml
[project]
name = "knn-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "isort>=8.0.1",
    "mypy>=1.20.2",
    "pandas>=3.0.2",
    "pandas-stubs>=3.0.0.260204",
    "pydocstyle>=6.3.0",
    "pyyaml>=6.0.3",
    "ruff>=0.15.11",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "types-pyyaml>=6.0.12.20260408",
]

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
pretty = true
show_error_codes = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = ["lgrs"]
lines_between_types = 1
lines_between_sections = 1
combine_as_imports = true
force_single_line = false
order_by_type = true
case_sensitive = false
skip = [".venv", "venv", "node_modules", "migrations", ".vscode"]
skip_gitignore = true
balanced_wrapping = true
atomic = true
